Struct interprocess_docfix::local_socket::LocalSocketStream

pub struct LocalSocketStream { /* private fields */ }

A local socket byte stream, obtained eiter from LocalSocketListener or by connecting to an existing local socket.

Examples

Basic client

Implementations

impl LocalSocketStream

pub fn connect<'a>(name: impl ToLocalSocketName<'a>) -> Result<Self>

Connects to a remote local socket server.

pub fn peer_pid(&self) -> Result<u32>

Retrieves the identifier of the process on the opposite end of the local socket connection.

Platform-specific behavior

macOS and iOS

Not supported by the OS, will always generate an error at runtime.

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<()>

Enables or disables the nonblocking mode for the stream. By default, it is disabled.

In nonblocking mode, reading and writing will immediately return with the WouldBlock error in situations when they would normally block for an uncontrolled amount of time. The specific situations are:

When reading is attempted and there is no new data available;

When writing is attempted and the buffer is full due to the other side not yet having read previously sent data.