Module interprocess_docfix::os::unix

Unix-specific functionality for various interprocess communication primitives, as well as Unix-specific ones.

FIFO files

This type of interprocess communication similar to unnamed pipes in that they are unidirectional byte channels which behave like files. The difference is that FIFO files are actual (pseudo)files on the filesystem and thus can be accessed by unrelated applications (one doesn’t need to be spawned by another).

FIFO files are available on all supported systems.

Unix domain sockets

Those are sockets used specifically for local IPC. They support bidirectional connections, identification by file path or inside the abstract Linux socket namespace, optional preservation of message boundaries (SOCK_DGRAM UDP-like interface) and transferring file descriptor ownership.

Unix domain sockets are not available on ARM Newlib, but are supported on all other Unix-like systems.

Modules

fifo_file

Creating and using FIFO files, which are also known as “named pipes” but totally different from named pipes on Windows.

signal

Signal support for Unix-like systems.

udsocket

Support for Unix domain sockets, abbreviated here as “Ud-sockets”.