Struct interprocess_docfix::os::unix::udsocket::tokio::UdStream

pub struct UdStream(/* private fields */);

A Unix domain socket byte stream, obtained either from UdStreamListener or by connecting to an existing server.

Examples

Basic client

Implementations

impl UdStream

pub async fn connect(path: impl ToUdSocketPath<'_>) -> Result<Self>

Connects to a Unix domain socket server at the specified path.

See ToUdSocketPath for an example of using various string types to specify socket paths.

pub fn split(&mut self) -> (BorrowedReadHalf<'_>, BorrowedWriteHalf<'_>)

Borrows a stream into a read half and a write half, which can be used to read and write the stream concurrently.

This method is more efficient than .into_split(), but the halves cannot be moved into independently spawned tasks.

pub fn into_split(self) -> (OwnedReadHalf, OwnedWriteHalf)

Splits a stream into a read half and a write half, which can be used to read and write the stream concurrently.

Unlike .split(), the owned halves can be moved to separate tasks, which comes at the cost of a heap allocation.

Dropping either half will shut it down. This is equivalent to calling .shutdown() on the stream with the corresponding argument.

pub fn reunite( read: OwnedReadHalf, write: OwnedWriteHalf ) -> Result<Self, ReuniteError>

Attempts to put two owned halves of a stream back together and recover the original stream. Succeeds only if the two halves originated from the same call to .into_split().

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of the stream. See Shutdown.

Attempting to call this method with the same how argument multiple times may return Ok(()) every time or it may return an error the second time it is called, depending on the platform. You must either avoid using the same value twice or ignore the error entirely.

pub fn get_peer_credentials(&self) -> Result<ucred>

Fetches the credentials of the other end of the connection without using ancillary data. The returned structure contains the process identifier, user identifier and group identifier of the peer.

pub fn into_sync(self) -> Result<SyncUdStream>

Detaches the async object from the Tokio runtime (therefore has to be called within the runtime) and converts it to a blocking one.

pub fn from_sync(sync: SyncUdStream) -> Result<Self>

Creates a Tokio-based async object from a blocking one. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error.

pub fn into_std(self) -> Result<StdUdStream>

Detaches the async object from the Tokio runtime and converts it to a blocking one from the standard library. Returns an error if called outside a Tokio runtime context.

pub fn from_std(std: StdUdStream) -> Result<Self>

Creates a Tokio-based async object from a blocking one from the standard library. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error.

pub fn into_tokio(self) -> TokioUdStream

Unwraps into Tokio’s corresponding type. This is a zero-cost operation.

pub fn from_tokio(tokio: TokioUdStream) -> Self

Wraps Tokio’s corresponding type. This is a zero-cost operation.

pub unsafe fn from_raw_fd(fd: c_int) -> Result<Self>

Creates a Tokio-based async object from a given raw file descriptor. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error (which is why the FromRawFd trait can’t be implemented instead).

Safety

The given file descriptor must be valid (i.e. refer to an existing kernel object) and must not be owned by any other file descriptor container. If this is not upheld, an arbitrary file descriptor will be closed when the returned object is dropped.

pub fn into_raw_fd(self) -> Result<c_int>

Releases ownership of the raw file descriptor, detaches the object from the Tokio runtime (therefore has to be called within the runtime) and returns the file descriptor as an integer.