Struct interprocess_docfix::os::unix::udsocket::tokio::UdStreamListener

pub struct UdStreamListener(/* private fields */);

A Tokio-based Unix domain byte stream socket server, listening for connections.

All such sockets have the SOCK_STREAM socket type; in other words, this is the Unix domain version of a TCP server.

Can be freely converted to and from its Tokio counterpart.

Examples

Basic server

Implementations

impl UdStreamListener

pub fn bind<'a>(path: impl ToUdSocketPath<'a>) -> Result<Self>

Creates a new listener socket at the specified address.

If the socket path exceeds the maximum socket path length (which includes the first 0 byte when using the socket namespace), an error is returned. Errors can also be produced for different reasons, i.e. errors should always be handled regardless of whether the path is known to be short enough or not.

Example

See ToUdSocketPath.

System calls

socket

bind

pub async fn accept(&self) -> Result<UdStream>

Listens for incoming connections to the socket, asynchronously waiting a client is connected.

pub fn into_sync(self) -> Result<SyncUdStreamListener>

Detaches the async object from the Tokio runtime (therefore has to be called within the runtime) and converts it to a blocking one.

pub fn from_sync(sync: SyncUdStreamListener) -> Result<Self>

Creates a Tokio-based async object from a blocking one. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error.

pub fn into_std(self) -> Result<StdUdStreamListener>

Detaches the async object from the Tokio runtime and converts it to a blocking one from the standard library. Returns an error if called outside a Tokio runtime context.

pub fn from_std(std: StdUdStreamListener) -> Result<Self>

Creates a Tokio-based async object from a blocking one from the standard library. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error.

pub fn into_tokio(self) -> TokioUdStreamListener

Unwraps into Tokio’s corresponding type. This is a zero-cost operation.

pub fn from_tokio(tokio: TokioUdStreamListener) -> Self

Wraps Tokio’s corresponding type. This is a zero-cost operation.

pub unsafe fn from_raw_fd(fd: c_int) -> Result<Self>

Creates a Tokio-based async object from a given raw file descriptor. This will also attach the object to the Tokio runtime this function is called in, so calling it outside a runtime will result in an error (which is why the FromRawFd trait can’t be implemented instead).

Safety

The given file descriptor must be valid (i.e. refer to an existing kernel object) and must not be owned by any other file descriptor container. If this is not upheld, an arbitrary file descriptor will be closed when the returned object is dropped.

pub fn into_raw_fd(self) -> Result<c_int>

Releases ownership of the raw file descriptor, detaches the object from the Tokio runtime (therefore has to be called within the runtime) and returns the file descriptor as an integer.