Struct interprocess_docfix::os::windows::named_pipe::DuplexMsgPipeStream

pub struct DuplexMsgPipeStream { /* private fields */ }

Message stream reader and writer for a named pipe.

Created either by using PipeListener or by connecting to a named pipe server.

Implementations

impl DuplexMsgPipeStream

pub fn connect(name: impl AsRef<OsStr>) -> Result<Self>

Connects to the specified named pipe (the \\.\pipe\ prefix is added automatically), blocking until a server instance is dispatched.

pub fn connect_to_remote( pipe_name: impl AsRef<OsStr>, hostname: impl AsRef<OsStr> ) -> Result<Self>

Connects to the specified named pipe at a remote computer (the \\<hostname>\pipe\ prefix is added automatically), blocking until a server instance is dispatched.

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<()>

Sets whether the nonblocking mode for the pipe stream is enabled. By default, it is disabled.

In nonblocking mode, attempts to read from the pipe when there is no data available or to write when the buffer has filled up because the receiving side did not read enough bytes in time will never block like they normally do. Instead, a WouldBlock error is immediately returned, allowing the thread to perform useful actions in the meantime.

If called on the server side, the flag will be set only for one stream instance. A listener creation option, nonblocking, and a similar method on the listener, set_nonblocking, can be used to set the mode in bulk for all current instances and future ones.

pub fn is_server(&self) -> bool

Returns true if the stream was created by a listener (server-side), false if it was created by connecting to a server (server-side).

pub fn is_client(&self) -> bool

Returns true if the stream was created by connecting to a server (client-side), false if it was created by a listener (server-side).

pub fn client_process_id(&self) -> Result<u32>

Retrieves the process identifier of the client side of the named pipe connection.

pub fn client_session_id(&self) -> Result<u32>

Retrieves the session identifier of the client side of the named pipe connection.

pub fn server_process_id(&self) -> Result<u32>

Retrieves the process identifier of the server side of the named pipe connection.

pub fn server_session_id(&self) -> Result<u32>

Retrieves the session identifier of the server side of the named pipe connection.

pub fn disconnect_without_flushing(self) -> Result<()>

Disconnects the named pipe stream without flushing buffers, causing all data in those buffers to be lost. This is much faster (and, in some case, the only finite-time way of ending things) than simply dropping the stream, since, for non-async named pipes, the Drop implementation flushes first.

Only makes sense for server-side pipes and will return an error if called on a client stream. For async pipe streams, this is the same as dropping the pipe.