Struct interprocess_docfix::os::windows::named_pipe::PipeListenerOptions

#[non_exhaustive]
pub struct PipeListenerOptions<'a> {
    pub name: Cow<'a, OsStr>,
    pub mode: PipeMode,
    pub nonblocking: bool,
    pub instance_limit: Option<NonZeroU8>,
    pub write_through: bool,
    pub accept_remote: bool,
    pub input_buffer_size_hint: usize,
    pub output_buffer_size_hint: usize,
    pub wait_timeout: NonZeroU32,
}

Allows for thorough customization of PipeListeners during creation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Cow<'a, OsStr>

Specifies the name for the named pipe. Since the name typically, but not always, is a string literal, an owned string does not need to be provided.

mode: PipeMode

Specifies how data is written into the data stream. This is required in all cases, regardless of whether the pipe is inbound, outbound or duplex, since this affects all data being written into the pipe, not just the data written by the server.

nonblocking: bool

Specifies whether nonblocking mode will be enabled for all stream instances upon creation. By default, it is disabled.

There are two ways in which the listener is affected by nonblocking mode:

Whenever accept is called or incoming is being iterated through, if there is no client currently attempting to connect to the named pipe server, the method will return immediately with the WouldBlock error instead of blocking until one arrives.

The streams created by accept and incoming behave similarly to how client-side streams behave in nonblocking mode. See the documentation for set_nonblocking for an explanation of the exact effects.

instance_limit: Option<NonZeroU8>

Specifies the maximum amount of instances of the pipe which can be created, i.e. how many clients can be communicated with at once. If set to 1, trying to create multiple instances at the same time will return an error. If set to None, no limit is applied. The value 255 is not allowed because of Windows limitations.

write_through: bool

Enables write-through mode, which applies only to network connections to the pipe. If enabled, writing to the pipe would always block until all data is delivered to the other end instead of piling up in the kernel’s network buffer until a certain amount of data accamulates or a certain period of time passes, which is when the system actually sends the contents of the buffer over the network.

Not required for pipes which are restricted to local connections only. If debug assertions are enabled, setting this parameter on a local-only pipe will cause a panic when the pipe is created; in release builds, creation will successfully complete without any errors and the flag will be completely ignored.

accept_remote: bool

Enables remote machines to connect to the named pipe over the network.

input_buffer_size_hint: usize

Specifies how big the input buffer should be. The system will automatically adjust this size to align it as required or clip it by the minimum or maximum buffer size.

output_buffer_size_hint: usize

Specifies how big the output buffer should be. The system will automatically adjust this size to align it as required or clip it by the minimum or maximum buffer size.

wait_timeout: NonZeroU32

The default timeout when waiting for a client to connect. Used unless another timeout is specified when waiting for a client.

Implementations

impl<'a> PipeListenerOptions<'a>

pub fn new() -> Self

Creates a new builder with default options.

pub fn to_owned(&self) -> PipeListenerOptions<'static>

Clones configuration options which are not owned by value and returns a copy of the original option table which is guaranteed not to borrow anything and thus ascribes to the 'static lifetime.

This is used instead of the ToOwned trait for backwards compatibility – this will be fixed in the next breaking release.

pub fn name(self, name: impl Into<Cow<'a, OsStr>>) -> Self

Sets the [name](#structfield. name ) parameter to the specified value.

pub fn mode(self, mode: impl Into<PipeMode>) -> Self

Sets the [mode](#structfield. mode ) parameter to the specified value.

pub fn nonblocking(self, nonblocking: impl Into<bool>) -> Self

Sets the [nonblocking](#structfield. nonblocking ) parameter to the specified value.

pub fn instance_limit( self, instance_limit: impl Into<Option<NonZeroU8>> ) -> Self

Sets the [instance_limit](#structfield. instance_limit ) parameter to the specified value.

pub fn write_through(self, write_through: impl Into<bool>) -> Self

Sets the [write_through](#structfield. write_through ) parameter to the specified value.

pub fn accept_remote(self, accept_remote: impl Into<bool>) -> Self

Sets the [accept_remote](#structfield. accept_remote ) parameter to the specified value.

pub fn input_buffer_size_hint( self, input_buffer_size_hint: impl Into<usize> ) -> Self

Sets the [input_buffer_size_hint](#structfield. input_buffer_size_hint ) parameter to the specified value.

pub fn output_buffer_size_hint( self, output_buffer_size_hint: impl Into<usize> ) -> Self

Sets the [output_buffer_size_hint](#structfield. output_buffer_size_hint ) parameter to the specified value.

pub fn wait_timeout(self, wait_timeout: impl Into<NonZeroU32>) -> Self

Sets the [wait_timeout](#structfield. wait_timeout ) parameter to the specified value.

pub fn create<Stream: PipeStream>(&self) -> Result<PipeListener<Stream>>

Creates the pipe listener from the builder. The Stream generic argument specifies the type of pipe stream that the listener will create, thus determining the direction of the pipe and its mode.

For outbound or duplex pipes, the mode parameter must agree with the Stream’s WRITE_MODE. Otherwise, the call will panic in debug builds or, in release builds, the WRITE_MODE will take priority.