Struct interprocess_docfix::unnamed_pipe::UnnamedPipeReader

pub struct UnnamedPipeReader { /* private fields */ }

A handle to the reading end of an unnamed pipe, created by the pipe function together with the writing end.

The core functionality is exposed in a file-like Read interface. On Windows, the ShareHandle and As-/Into-/FromRawHandle traits are also implemented, along with As-/Into-/FromRawFd on Unix.